+++ bin/index.js
@@ -195,9 +195,12 @@ async function appUsageMenu() {
           "Show last week",
           "Show last month",
           "Show last year",
+          "Show specific day",
+          "Show date range",
           "Help",
           "Back",
         ],
+        pageSize: 15,
       },
     ]);
 
@@ -249,6 +252,108 @@ async function appUsageMenu() {
         defaultTable.addRows(results);
 
         // Clear the console and print the table
+        console.clear();
+        console.log(SMALL_SEPARATOR);
+        defaultTable.printTable();
+        break;
+
+      case "Show specific day":
+        const { specificDate } = await inquirer.prompt([
+          {
+            type: "input",
+            name: "specificDate",
+            message: "Enter the date in the format (DD MM YYYY):",
+            validate: (input) => {
+              const isValidDate = /^\d{2} \d{2} \d{4}$/.test(input.trim());
+              return isValidDate
+                ? true
+                : "Please enter a valid date in the format DD MM YYYY. For example, 21 04 2024.";
+            },
+          },
+        ]);
+
+        // Convert the space-separated date into a Date object
+        const [day, month, year] = specificDate.split(" ").map(Number);
+        const formattedDate = `${String(day).padStart(2, "0")}/${String(
+          month
+        ).padStart(2, "0")}/${year}`;
+        const displayDate = new Date(year, month - 1, day).toDateString();
+
+
+        // Fetch the app usage data for the specific day
+        results = await fetchAppUsage(formattedDate);
+
+        // Update and print the table
+        defaultTable.table.title = `Apps usage report for: ${displayDate}`;
+        defaultTable.addRows(results);
+        console.clear();
+        console.log(SMALL_SEPARATOR);
+        defaultTable.printTable();
+        break;
+
+      case "Show date range":
+        const { startDate, endDate } = await inquirer.prompt([
+          {
+            type: "input",
+            name: "startDate",
+            message: "Enter the start date (DD MM YYYY):",
+            validate: (input) => {
+              const isValidDate = /^\d{2} \d{2} \d{4}$/.test(input.trim());
+              return isValidDate
+                ? true
+                : "Please enter a valid date in the format DD MM YYYY. For example, 01 01 2024.";
+            },
+          },
+          {
+            type: "input",
+            name: "endDate",
+            message: "Enter the end date (DD MM YYYY):",
+            validate: (input) => {
+              const isValidDate = /^\d{2} \d{2} \d{4}$/.test(input.trim());
+              return isValidDate
+                ? true
+                : "Please enter a valid date in the format DD MM YYYY. For example, 31 12 2024.";
+            },
+          },
+        ]);
+
+        // Convert the space-separated dates into Date objects
+        const [startDay, startMonth, startYear] = startDate
+          .split(" ")
+          .map(Number);
+        const [endDay, endMonth, endYear] = endDate.split(" ").map(Number);
+
+        const formattedStartDate = `${String(startDay).padStart(
+          2,
+          "0"
+        )}/${String(startMonth).padStart(2, "0")}/${startYear}`;
+        const formattedEndDate = `${String(endDay).padStart(2, "0")}/${String(
+          endMonth
+        ).padStart(2, "0")}/${endYear}`;
+
+        const displayStartDate = new Date(
+          startYear,
+          startMonth - 1,
+          startDay
+        ).toDateString();
+        const displayEndDate = new Date(
+          endYear,
+          endMonth - 1,
+          endDay
+        ).toDateString();
+
+        // Fetch the app usage data for the specified date range
+        results = await fetchAppUsage(
+          {
+            startDate: formattedStartDate,
+            endDate: formattedEndDate,
+          },
+          true
+        );
+
+        // Update and print the table
+        defaultTable.table.title = `Apps usage: \nFrom: ${displayStartDate} -> ${displayEndDate}`;
+        defaultTable.addRows(results);
         console.clear();
         console.log(SMALL_SEPARATOR);
         defaultTable.printTable();