+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## v1.1.0 
+- Added the possibility to get the report of apps usage for a specific date
+- Added the possibility to get the report of apps usage for a specific date range
+
+
 ## v1.0.0 Initial Release
 - Initial release of the package